+++ plom_server/Scan/views/scanner_home.py
@@ -34,11 +34,10 @@ from plom.plom_exceptions import PlomBundleLockedException
 class ScannerOverview(ScannerRequiredView):
     def get(self, request: HttpRequest) -> HttpResponse:
         context = self.build_context()
+        total_papers = ManageScanService.get_total_papers()
+        completed_papers = ManageScanService.get_number_completed_papers()
+        incomplete_papers = ManageScanService.get_number_incomplete_papers()
         mss = ManageScanService()
-
-        total_papers = mss.get_total_papers()
-        completed_papers = mss.get_number_completed_papers()
-        incomplete_papers = mss.get_number_incomplete_papers()
         pushed_bundles = mss.get_number_pushed_bundles()
         unpushed_bundles = mss.get_number_unpushed_bundles()
         discards = mss.get_discarded_page_info()